requirements.txt
mcp>=0.1.1
httpx>=0.27.0
PyYAML>=6.0.1
pytest>=8.3.3
pytest-asyncio>=0.24.0
